+++ Сниппеты/[Даты, Время]/Таймер пройденого времени в UnixTime.cs
@@ -1,7 +1,7 @@
 // ТАЙМЕР ПРОЙДЕННОГО ВРЕМЕНИ В UNIXTIME
 
 // Стартовое время в UnixTime
-long startTime  = int.Parse(project.Variables["startTime"].Value);
+long startTime  = long.Parse(project.Variables["startTime"].Value);
 
 // Время сейчас в UnixTime
 long unixTimeNow = ((DateTimeOffset)DateTime.UtcNow).ToUnixTimeSeconds();
